// apps/web/src/shared/api/generated/model/publishersControllerLibraryListGeography.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * book-nest API
 * REST API for the book-nest project
 * OpenAPI spec version: 1.0
 */

export type PublishersControllerLibraryListGeography =
  (typeof PublishersControllerLibraryListGeography)[keyof typeof PublishersControllerLibraryListGeography];

export const PublishersControllerLibraryListGeography = {
  all: "all",
  ua: "ua",
  foreign: "foreign",
  unknown: "unknown",
} as const;