// apps/web/src/shared/api/generated/model/publishersControllerLibraryListOrder.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * book-nest API
 * REST API for the book-nest project
 * OpenAPI spec version: 1.0
 */

export type PublishersControllerLibraryListOrder =
  (typeof PublishersControllerLibraryListOrder)[keyof typeof PublishersControllerLibraryListOrder];

export const PublishersControllerLibraryListOrder = {
  asc: "asc",
  desc: "desc",
} as const;