// apps/web/src/shared/api/generated/model/publishersControllerLibraryListSort.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * book-nest API
 * REST API for the book-nest project
 * OpenAPI spec version: 1.0
 */

export type PublishersControllerLibraryListSort =
  (typeof PublishersControllerLibraryListSort)[keyof typeof PublishersControllerLibraryListSort];

export const PublishersControllerLibraryListSort = {
  name: "name",
  booksCount: "booksCount",
  readCount: "readCount",
  wantToBuyCount: "wantToBuyCount",
  averageRating: "averageRating",
  lastBookAddedAt: "lastBookAddedAt",
} as const;