// apps/web/src/shared/api/generated/model/publishersControllerLibraryListSource.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * book-nest API
 * REST API for the book-nest project
 * OpenAPI spec version: 1.0
 */

export type PublishersControllerLibraryListSource =
  (typeof PublishersControllerLibraryListSource)[keyof typeof PublishersControllerLibraryListSource];

export const PublishersControllerLibraryListSource = {
  all: "all",
  global: "global",
  custom: "custom",
} as const;